requirements.txt
backports.zoneinfo==0.2.1
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dnspython==2.1.0
et-xmlfile==1.1.0
exchangelib==4.4.0
idna==2.10
isodate==0.6.0
lxml==4.6.3
ntlm-auth==1.5.0
numpy==1.21.0
oauthlib==3.1.1
openpyxl==3.0.7
pandas==1.2.5
pycparser==2.20
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
six==1.16.0
tzdata==2021.1
tzlocal==2.1
urllib3==1.26.6
